Move the call to insertReferenceTemps() as late as it can go in the translation. [reviewed by kbrady]
Split the insertReferenceTemps() function into two.  The expansion of
INIT_FIELDS primitives was blended in with the original.
insertReferenceTemps() was moved from functionResolution.cpp to
callDestructors.cpp, and the call to it was placed at the end of the main
function for the callDestructors pass.

It is undesirable to insert reference temps early in translation, because the
inserted ADDR_OF primitives may have to be handled specially in subsequent
passes.  The immediate motivation for moving insertReferenceTemps() to the end
of callDestructors() is that the special case of ADDR_OF comes up in the
implementation of returnRecordsByReferenceArguments(), which is called earlier
in the callDestructors pass.  Moving insertReferenceTemps() after that call
obviates the special handling.

In might be possible to push the call to insertReferenceTemps() into the next
pass (lowerIterators), but it would have to appear early in that pass.
lowerIterators calls localCopyPropagation(), and copy propagation relies on the
reference temps having been inserted into the AST.  Also, insertReferencTemps()
is somewhat related to insertYieldTemps() which also appears in the
callDestructors pass.  So callDestructors seems like a better place for it.

[tested on / (std)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22632 3a8e244f-b0f2-452b-bcba-4c88e055c3ca