Change extern string arguments to be of type c_string
[reviewed by Kyle]

This patch changes the signature of extern string and qio procedures
to take c_string instead of Chapel string.  Functionally, nothing has
changed as c_string and Chapel string are still the same in the
generated code, but this will be useful for Kyle now, so I peeled it
out of the other string work.

chpltypes.[ch]:

- Change chpl_string to c_string.  I did bother making the
  corresponding change in the modules to use s.c_str(), since it does
  not matter now and might be different in the future.

modules/standard/IO.chpl:

- Change Chapel string to c_string for extern prototypes in most
  cases.  I omitted qio_channel_read_string() and
  qio_channel_scan_string(), because here appears to be a problem with
  codegen of ref c_string.  Kyle will hopefully be fixing this case.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22638 3a8e244f-b0f2-452b-bcba-4c88e055c3ca