Remove default case from switch statement in intentDescrString
[reviewed by bradc]

Related to r22621 and r22618:

This removes the default case from the switch statement inside of
intentDescrString so the compiler will give a warning that all cases are not
handled if IntentTag is modified. An INT_FATAL is also put in to catch when a
case was not handled in the switch statement.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22639 3a8e244f-b0f2-452b-bcba-4c88e055c3ca