Flip default value of dataParIgnoreRunningTasks to false [reviewed by Sung]
It was discovered that our nested parallelism was not being handled as well
as we would like, namely that programs were greatly slowed down when it
occurred due to the number of tasks created and the division of labor that went
with their creation.  As a short-term solution, we decided to change the value
of dataParIgnoreRunningTasks to false, so that we would make best use of the
threads we had already created.  This showed a drastic improvement in the
performance of the fifo tasking layer.  However, it is a short-term solution
and more work will need to be done, both for fifo and for the other tasking
layers.

This commit also includes alterations to tests that are affected by the value
changing.  A couple of tests set this variable back to true, the rest were
trivial.

Tested on std, gasnet-everything, and performance for fifo.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22641 3a8e244f-b0f2-452b-bcba-4c88e055c3ca