Move loop-invariant code out of a loop in computeLoopInvariants(). [reviewed by eronaghan]
Since the potential aliases due to formal arguments potentially referencing the
same memory can be computed once per function (as opposed to once per basic
block), the loop to perform this search can be moved out of the basic block
traversal (and was).  

This patch also contains one typo correction and an indentation adjustment.

[smoke-tested with jacobi.chpl]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22648 3a8e244f-b0f2-452b-bcba-4c88e055c3ca