Future for begin statements in initCopy and autoCopy calls
Due to the implicit sync around main, all begin statements require the
global _endCount var for this sync, which requires it to be passed
through all enclosing calls.

The removeUnnecessaryAutoCopyCalls pass asserts if autoCopy and
initCopy calls if the call has more than 1 argument.

Note that the .bad file for autoCopyWithBegin.chpl will not match when
CHPL_VERIFY is set, as it gets caught by the verification pass.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22664 3a8e244f-b0f2-452b-bcba-4c88e055c3ca