Un-futurized and fixed up ref-intent-bug-2big.chpl.
At some point this test stopped hitting segfault under gasnet,
and instead started hitting an assertion in the test itself.

... which was rightfully so, because I had not realized that
'coforall' does not go to remote locales the way 'forall' does.
So my "sanity" assertion was catching that.

... and even without that, if targetLocales has 'nil' for
all its elements, there would be no remote locales to start with.

It is surprising that Block constructor can be invoked with
targetLocales filled with nils! I investigated this issue separately.

... and even it if weren't filled with nills, the .good file would not match.
That's because of the following:

The .prediff for this test sorts its output. 'sort' sorts differently
depending on the value of LC_ALL (or LC_COLLATE). I adjusted .good
for the sorting order with LC_ALL unset (r20926). However,
start_test actually sets LC_ALL=C, 'sort' in .prediff inherits that,
and we get a mismatch.

Therefore:

* I am un-futurizing this test because it does not exhibit any issues
with the Chapel implementation.

* I am fixing the test source to populate targetLocales (called 'gridLocales')
as intended.

* I am fixing the test source to do 'on' within the 'coforall'.
These two fixes make the test work as expected.

* I am reverting .good to be sorted with LC_ALL=C, to match the output
of the .prediff.

* I am removing .skipif because the test works fine for comm=none.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22681 3a8e244f-b0f2-452b-bcba-4c88e055c3ca