This adds a new flag to the compiler called --print-passes-file <filename>
[reviewed by tom]

It is basically the same thing as print-passes but instead of just printing to
stdout this prints to the file. This is to be used for compiler performance
testing where you don't want other compiler output to get in the way or for
--print-passes to mess up expected compiler output. 

This does not override print-passes and the two can be used at the same time
(thought I can't imagine when you would need that.) If the specified file
cannot be opened, an error message is displayed but no recovery attempt is
made. 

The commit also updates the test checking for changes to help and updates the
man page. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22683 3a8e244f-b0f2-452b-bcba-4c88e055c3ca