Changes to performance graphs for compiler performance
[reviewed by sung]

This commit adds two main features to the performance graphs. The first is that
the series in each graph will now be sorted based on data from the most recent
date. So series with the highest values will be on top in the graph. The second
is that graphs can now be expanded. This is primarily for the compiler
performance graphs but it might be nice to add an expand button. Expand takes a
graph with multiple series and turns that into a graph with all the series
still and then a graph for each series that shows that series only. For
instance if you have all (currently 42) compiler passes timings it is too much
information to see them all at once or to single one out. Expansion will create
1 graph with all 42 and 42 graphs with each individual series. Expansion does
not have to expand all. For instance expansion of 10 will only expand 10 series
and the original graph will only have the first 10 series displayed. 

The other changes to this are largely some refactoring. The callbacks have been
pulled out of the genDygraph method. 

In addition the log button will now only appear after the graph is fully
rendered. There was previously a race condition where if you loaded multiple
graphs and tried to zoom or pan before all the graphs were fully loaded an
attempt was made to change the zoom on graphs that didn't exist yet which led
to some odd crashes. Beyond that we were leaking a lot of memory because graphs
were never fully destroyed. This was fixed by calling dygraphs destroy method
when we remove graphs. The dygraphs date parser is now being used to calculate
time to avoid weird issues with daylight savings that occured with Date.parse.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22685 3a8e244f-b0f2-452b-bcba-4c88e055c3ca