Update no-local and gasnet .good files with reduced comm counts
My r22638 had the pleasant side effect of reducing the number of times
we widen a string via chpl_widen_string as well as call
chpl_comm_wide_get_string.  This is the only mem counts test that
wraps a call to writeln, which is why the leaked memory count went
down for the 'Total' count.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22689 3a8e244f-b0f2-452b-bcba-4c88e055c3ca