Change prototype of qio_channel_print_literal() back to using a Chapel string
[reviewed by Kyle]

Following up on r22638, qio_channel_print_literal() causes issues
during codegen to extern functions, because the generated code can
pass a non-local Chapel string to c_string_from_wide_string().

So for now, we are going to change it back to use a Chapel string
until we sort out issues with codegen and strings in general.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22690 3a8e244f-b0f2-452b-bcba-4c88e055c3ca