Retire some futures which are passing due to assign-to-ref (r22674). [trivial]
These tests were futures because alias tracking in the copyPropagation
optimization is somewhat primitive.  It will invalidate a candidate pair if an
alias is created from a member of that pair, but only if the alias creation
follows the creation of the candidate pair.  To be strictly correct, however,
aliases must be tracked across an entire function, and the creation of
a candidate pair prevented if either member is already aliased.

In these tests, the failure was due to refPropagation removing alias creation
statements that followed the creation of certain candidate pairs.  The
refPropagation pass formerly preceded the copyPropagation pass.  As a result,
copyPropagation would make a substitution that was invalid.

In the tests, the invalid substitution manifests as a an expected update
(through the alias) that fails to occur when the program is run.

The reason why r22674 cured the problem is that (as a workaround for other
failures), I reversed the order of the refPropagation and copyPropagation
passes.  As a result, the alias creation statements are in the "right place"
for the sake of copyPropagation.  The behavior of the aliases is preserved, and
the program runs as expected.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22713 3a8e244f-b0f2-452b-bcba-4c88e055c3ca