Add error messages to the timeout from in paratest.server. [reviewed by vass]
Updated collect_logs to print a general error message and note which workers
timed out if paratest.server terminates before getting a response from all
worker nodes.  These messages are also echoed to stdout.

The wait in free_workers was removed, because it interferes with the timeout
code.  Now you will notice that comma is printed each second after the last
directory is doled out until all workers have reported in (or paratest.server
times out).  Previous behavior would wait for a child process to die, which
saved paratest.server a bit of processing time at the very end (but who
cares?).

The $done_or_dead count was split so that we can check the $dead count at the
end of the run and print out the names of nodes that died before reporting in.

The clause containing
       print "\nExiting early due to failures on all nodes\n";
was removed.  $nodeCount has be redefined so it is always equal to the initial
number of workers.  So the clause will no longer fire. Ever.

[tested on examples/]

 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22715 3a8e244f-b0f2-452b-bcba-4c88e055c3ca