Filing futures against bad zippering
Kyle's retirement of some futures yesterday that I've traditionally
attributed to an inherent flaw in the leader-follower 1.0 interface
made me wonder how they could be fixed when we haven't fixed the flaw.
It turns out it's because array assignment gives us a special place to
catch such mismatches.  For a pure zippered iteration, the bug still
exists, so I'm capturing it here in three tests (arrayZipMismatch,
arrayZipMismatch2, and rangeSizeMismatch).  For simplicity, I've given
them all the same .future though arguably arrayZipMismatch2.chpl is
more of an error message future than a bug -- it is correct that it
doesn't work, it just doesn't report the issue in a helpful way.

While here, I accidentally found myself forgetting the zip() keyword
and getting some surprising behavior suggesting that we're permitting
arrays to be detupled in an iterator context (I can't seem to
reproduce it in a non-iterator context), so I'm filing a bug against
that as well.  This seems conceptually related to the tuple size
mismatch bug that David fixed in the past week or so, but probably
stems from a different place in the code.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22725 3a8e244f-b0f2-452b-bcba-4c88e055c3ca