Remove two passing .future files
users/ferguson/exit users/ferguson/forall_expr both pass for qthreads+gasnet
and qthreads+none.  The .future for forall_expr said it was showing an
intermittent problem with qthreads, but I ran it successfully 1000 times.
Also removed the .skipif files that were restricting the tests to only run for
qthreads.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22728 3a8e244f-b0f2-452b-bcba-4c88e055c3ca