Update comm counts in .good files due to r22726
Update the comm counts in the .good files for test that measure
performance of parallel iteration using BlockDist.

The number of gets from locales != 0 is reduced by one for each
invocation of the BlockDom iterator, because we no longer use
locDom.myBlock to compute the number of tasks as it is now done in the
DefaultRectangularDom iterator.  We could have been using the local
tmpBlock there in the version before forwarding, but just never
noticed.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22729 3a8e244f-b0f2-452b-bcba-4c88e055c3ca