Revise check for loop block primitives in optimizeOnClauses. [reviewed by sungeun]
The test users/ferguson/ontest.chpl was failing due to the assertion in
isFastPrimitive [optimizeOnClauses] that loop block primitives do not remain in
the tree when optimizeOnClauses run.  

The assertion is false in general (a quick scan of the AST reveals many
instances).  But since isFastPrimitive() is invoked only on the calls in a
function that is a candidate for fast-forking (or a function called by the
candidate, recursively), most tests were getting past this assertion.  

Although ontest.chpl started passing after the assign-to-ref patch was
committed [r22674], the root cause of the original failure has not been
addressed.  This patch is intended to ameliorate that situation (so the
assertion failure does not magically reappear at a later date).


DETAILS

    compiler/optimizations/optimizeOnClauses.cpp
Loop block primitives were moved out of the "no longer in tree" assertion in
isFastPrimitive().  They now return 'true' because loops in and of themselves
do not involve task control or synchronization.

    compiler/main/checks.cpp
Added a new check for primitives that should be removed by the
createTaskFunctions pass.  Note that these do not include the loop block
primitives.  Inserted the test in appropriate places in the verification
scaffold.

    compiler/AST/expr.cpp
Removed nonstandard parentheses from case labels in CallExpr::verify().


[tested with --verify on examples/]
[tested with gasnet on domains/ distributions/ multilocale/ performance/]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22730 3a8e244f-b0f2-452b-bcba-4c88e055c3ca