Retire ontest.future. [not reviewed]
Following r22674, this future started passing.

The original fault involved an assertion failure (actually INT_FATAL) in
isFastPrimitive() [optimizeOnClauses.cpp].  The assertion was that loop block
primitives such as PRIM_BLOCK_WHILEDO_LOOP never appear in the tree that late
in compilation.  But the fact is that they do, as Michael's test case
discovered.  The reason why most tests pass is that isFastPrimitive() is called
selectively, rather than for every call in the tree.

The "correct" solution is to treat those loop block primitives as valid at that
point in the compilation.  That change was made in r22730.  Based on that
change, it is appropriate to retire ontest.future and the associated .bad file.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22737 3a8e244f-b0f2-452b-bcba-4c88e055c3ca