Added the "performance label" feature and -numlocales to start_test and sub_test.
[reviewed by Kyle]

We would like to be able to specify distinct configurations of a test
(compiler options, number of locales, etc.) for different testing
configurations (e.g. an XC vs. a cluster vs. a desktop).

This commit allows doing so by allowing distinct "labels" in configuration
files' names/suffixes. By default, upon "start_test -performance ...",
the label is "perf" and these files are consulted:

   testname.perfkeys
   testname.perfcompopts
   PERFCOMPOPTS
   etc.

This can be overwritten with the "-perflabel" option.
For example, upon "start_test -performance -perflabel multiloc- ...",
these files are consulted:

   testname.multiloc-keys
   testname.multiloc-compopts
   MULTILOC-COMPOPTS
   etc.

Without the -performance option, the label is not consulted; or somewhat
equivalently the label is the empty string.

As a side effect, now NUMLOCALES/.numlocales files are not consulted
for performance testing. Instead, PERFNUMLOCALES/.perfnumlocales are
(when they exist).

The -numlocales option sets the number of locales that is used
in the absence of an applicable numlocales/NUMLOCALES file.

IMPLEMENTATION DETAILS

* The variable CHPL_TEST_PERF_LABEL conveys the performance label.
It is set only when -performance.

* I replaced references to the ".perfkeys" string with ".$perfkeys"
in start_test. Except I kept .perfkeys for spec test stats.

In sub_test:

* I defined PerfSfx(), PerfDirFile(), PerfTFile() to add the performance label
to the file names/suffixes, instead of the hard-coded "perf" label.
Used those functions appropriately.

* I moved the check for $CHPL_TEST_PERF earlier because a few things
depend on it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22742 3a8e244f-b0f2-452b-bcba-4c88e055c3ca