Fix gasnet performance regressions due to r22674. [reviewed by kbrady]
Changing the signature of the internally-generated record- and tuple-assignment
operators from 
  proc =(a, b) : a.type
to
  proc =(ref a, b) : void
causes one fewer get, but also entails a put for each field in the
structure.  Prior to the change, the returned value was copied back using a
bulk put, so for structures containing more than one field, the put count has
gone up.

The fix provided here is to hand-optimize the assignment operator.  The
user-defined assignment operator uses a bulk copy (PRIM_ASSIGN) in place of the
field-by-field assigment provided by the compiler by default.  This results in
a reduction of the number of gets for the three tests involved.  It also
resulted in a reduction in the number of puts for the small tuple case by
overriding the special case in the backend (PRIM_MOVE) for copying small
tuples.

[these specific tests tested with CHPL_COMM=gasnet (everything)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22744 3a8e244f-b0f2-452b-bcba-4c88e055c3ca