Make variables used to synchronize pthreads volatile.
[Proposed by David, reviewed by Sung and Greg.]

The chpl_qthread_done_initializing and done_finalizing flag variables
are used to synchronize pthreads, and therefore need to be volatile.
The details of their current usage cause this not to matter right now,
but those details may not always hold and even if they did, it's good
coding style to make them volatile.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22754 3a8e244f-b0f2-452b-bcba-4c88e055c3ca