Disable building of shared libraries for qthreads and hwloc, and fix static linking with CHPL_TASKS=qthreads
[reviewed by Greg, discussed with Dylan]

This patch disables shared libraries for both the qthreads and hwloc
packages, as it seems the safest way to make sure things work out of
the box.  That said, we need to throw LDFLAGS=-static for the hwloc
config for building some of thei binaries, which is not recognized for
all compilers.  Specifically, -static is not recognized directly by
Cray (craycc) or PGI (pgcc), tho it is recognized by the Cray PE
wrappers.  Given that we don't test these compilers in this
configuration and don't know off-hand anyone who does build in such a
way, we've opted to let this go and fix it later if needed.  In the
meantime, we will document it in the currently non-existent README
that someone needs to write for the hwloc directory.

In addition, in order to fix --static linking of Chapel programs, for
hwloc config I had to disable the use of libxml2 as well as disable
PCI device discovery, both of which Dylan confirmed are not needed for
qthreads.  Finally, I added a -lpthread and -rt to the link line when
compiling programs for CHPL_TASKS=qthreads.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22756 3a8e244f-b0f2-452b-bcba-4c88e055c3ca