Have sub_test issue a warning when encountering an empty .compopts file.
[reviewed by Lydia]

Previously, when sub_test encounters an empty .compopts file, it
silently skipped the test.

Indeed, sub_test reads a .compopts file as follows:

  compoptslist = ReadFileWithComments(execname+compoptssuffix, False)

and then iterates over compoptslist:

  for compopts in compoptslist:
      ... handle one .compopts entry: compile, execute, etc. ...

If .compopts is empty, 'compoptslist' is an empty list, so the 'for' loop 
executes 0 iterations, meaning the test does not get tested.

I added a check whether 'compoptslist' is an empty list.
If so, it issues a warning and adds a trivial element in that list,
just likewhat is done if there is no .compopts file at all.

Note: this adjustment (list is empty => add a trivial element)
is already done for execoptslist, except it does not issue a warning there.
Maybe it should?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22758 3a8e244f-b0f2-452b-bcba-4c88e055c3ca