Remove weird_bbox_block_works.* and add newly failing cases to weird_bbox_block.execopts
The Block leader forwarding change (r22726) changed the range used for
the computeBlock() calculation, so now it causes overflow in cases
that used to work.

I've moved the execopts that were teased out of the original test back
into weird_bbox_block.execopts and removed the "works" version, as it
clearly does not.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22759 3a8e244f-b0f2-452b-bcba-4c88e055c3ca