Update filterSuppressions to look for word boundaries around each suppression
[Reviewed by Elliot]

filterSuppressions was using a string.find to check for any substring that
matched a suppression. If one suppression was a substring of another and the
smaller one was first in the list, the longer suppression would never be found.
This would causing an error message to be printed.

Now a regex search is used to ensure the suppression is bordered by word
boundaries before it is returned as a match.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22764 3a8e244f-b0f2-452b-bcba-4c88e055c3ca