Remove accelerator locale model
[reviewed by greg]

This essentially reverts revisions 22384 and 22385. 

The accelerator locale model was added with the expectation that it would be
updated and polished before the 1.9 release. Since accelerator support
requiring the locale model is no longer going to be part of the release this
removes that code to avoid confusion and clutter. Setting the accelerator
locale model would have no effect for the release and it seems better to remove
the prototype than have it left in non-functional. 

It will be easy to revert this patch and add the accelerator locale model back
in when it is needed again. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22770 3a8e244f-b0f2-452b-bcba-4c88e055c3ca