Fix hwloc and qthreads builds on darwin
[Reviewed by gbt]

r22756 changed a few options on hwloc and qthreads builds that are incompatible
with darwin. This patch makes it so the options are only set on non-darwin
builds.

runtime/etc/Makefile.tasks-qthreads:
librt doesn't exist in darwin

third-party/hwloc/Makefile:
-static causes autoconf to attempt to create a statically linked binary to test
if the C compiler works... statically linked binaries are not supported under
darwin so it fails.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22772 3a8e244f-b0f2-452b-bcba-4c88e055c3ca