Fix problem with link order with -llvm
[reviewed by Greg]

Tiffany Mintz pointed out a problem using -l libraries
when compiling with -llvm. Turns out we were adding the
libraries in the wrong part of the link line - they should
be at the end, after the .o files. This patch fixes the
problem and makes the clang++ link line more similar to
the one in the C backend's Makefiles.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22776 3a8e244f-b0f2-452b-bcba-4c88e055c3ca