fix portability problems in test/ sh and csh scripts
[reviewed by Brad]

I recently switched to using Debian 7.4 for my Chapel development,
and I discovered:
- /bin/sh is more POSIX-compilant and doesn't like bash-isms
- /bin/csh is not tcsh and behaves oddly with echo >> somefile
  (namely that echo >> somefile does not add a newline, but
   echo "" >> somefile does).

The main /bin/sh problem is that [ "a" == "b" ] does not
work; what we want is [ "a" = "b" ] for string comparison
or [ 1 -eq 2 ] for numeric comparison.

The /bin/csh problems caused me to convert two tests
to using /bin/sh :
 test/memory/shannon/outofmemory/sub_test
and lots of tiny scripts in
 test/compflags/bradc/printstuff/

since the group has been trying to move away from csh. 




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22780 3a8e244f-b0f2-452b-bcba-4c88e055c3ca