Make dataParIgnoreRunningTasks=true by default for numa iteration and also make it obey dataParTasksPerLocale
[previewed by David, will be reviewed post-commit]

This commit makes the default for dataparIgnoreRunningTasks to be
true, because we don't (yet) have a way to query the number of running
tasks on a per sublocale basis.

This commit also fixes numa iteration over DefaultRectangular and
RangeBase to consider dataParTasksPerLocale.  Similarly, this is also
bit of a temporary solution, as we would really like to be able to
query for defaults for these values on a per sublocale basis.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22782 3a8e244f-b0f2-452b-bcba-4c88e055c3ca