Making --fast and --verify testing skip futures
[minimal, not verified]

For tests that aren't working, we tend to get different behavior when
compiling with --fast and --verify, and it arguably just wastes time.
This commit runs these test configurations with --no-futures in order
to save time.  Once the future is retired, if it starts failing
specifically in these modes, that seems like the proper time to deal
with it anyway.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22784 3a8e244f-b0f2-452b-bcba-4c88e055c3ca