Add performance futures suggesting bulk copy optimization to replace field-by-field remote puts. [not reviewed]
The specification states that the compiler-supplied record (and tuple)
assignment operators perform a field-by-field assignment.  But if these
assignments are also compiler-generated down to fundamental types, then the
entire assignment is equivalent to a bulk copy.

Right now, updating the fields in a remote object is expensive, because each
field write entails a remote put.  Replacing that with a bulk copy (where it is
legal) will replace those multiple remote puts with a single one.

[tested by themselves (gasnet-everything)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22786 3a8e244f-b0f2-452b-bcba-4c88e055c3ca