Removing --fast and fixing a bug
[trivial, not reviewed]

As a background activity, I've been looking at tests and directories
that use the --fast flag for correctness testing, which we've
generally stated we want to shy away from for fear of hiding bugs.
This directory contained an example in which --fast wasn't required
for correctness and was hiding an issue where a shorter range was
zippered against a longer one in the tests.  I addressed this by using
the # operator on the second range, cutting it down to the size of the
first.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22790 3a8e244f-b0f2-452b-bcba-4c88e055c3ca