Replaced an assertion in inlineFunctions.cpp. [reviewed by eronagha]
This is part of the copyPropagation rework.  Elliot recommended that I check it
in separately.

An assertion in inlineFunctions fires if one of the formal arguments in the
*containing* function appears in the inlined function's return statement
(before the latter is inlined).  The rework of copy propagation causes the
assertion to fire (i.e. fail), but there's actually nothing wrong with that.

The original assertion was intended to prevent a *formal* argument of the
*inlined* function from appearing in a return.  This makes sense, since the
formals should all be replaced by their corresponding actuals after the
function is inlined.

I put in a new assertion to this effect, but one can make static arguments that
the assertion will never fire, so it can also be safely removed as dead code.

[smoke-tested on hello.chpl (std)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22794 3a8e244f-b0f2-452b-bcba-4c88e055c3ca