Get rid of unneeded special case in code generation. [reviewed by eronagha]
While working on copy propagation, I noticed that there is a case in code
generation for PRIM_MOVE that special-cases the handling of RHS symbols that
have FLAG_STAR_TUPLE or FLAG_FIXED_STRING.  However, under the asssumption that
the RHS is never passed by ref, the actual code generation devolves to the
immediately-following general case.

I added an assert that the RHS is never passed by ref and ran full test run.
This proves that the assertion is true, so the entire clause can be removed.

I removed the clause and ran full (std) testing again.  No regressions were
reported.

[tested on test/ (std)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22795 3a8e244f-b0f2-452b-bcba-4c88e055c3ca