Re-implement copy propagation to handle aliases better. [reviewed by eronaghan]
As a workaround while adding support for assignment to ref [r22674], the order
of the refPropagation and copyPropagation passes was reversed.  This was
because the existing implementation of copy propagation does not do a good job
of tracking aliases.  The increased use of temporary variables of reference
type (due to using the new signature for assignment, passing the LHS argument
by reference) allowed refPropagation to make substitutions in more place.  And
that exposed the weakness in the current implementation of copy propagation.

The new implementation tracks aliases on a function-by-function basis.  By
definition, copy propagation may substitute one variable for another, but only
under the assumption that neither has changed since the pair was created.
Alias tracking allows us to determine when a variable may have been updated
indirectly; in that case candidate pairs involving that variable are
invalidated.

Aside from that, the algorithms for both local and global copy propagation are
taken straight from Muchnick.  The implementation of the predicate functions
isDef(), isUse(), derefUse(), and isRefUse() are used in place of the existing
isDefAndOrUse() to allow greater precision.  They were mostly created from
specific knowledge of the semantics of the primitives involved, with their
correctness verified experimentally.

The refPropagation pass was removed as a test (since it should be redundant
with copyPropagation now.  However, code size increased, so it was left in
place.

Similarly, since GCP uses LCP as an initial step, it was considered possible to
use only GCP.  However, this created performance problems on some larger
tests.  To many candidate pairs were getting created and stored.  Since the
invalidation of candidate pairs is quadratic (in the current implementation),
keeping the total number of candidate pairs low is a big win.


TODO
 - Add a check to ensure that a reference variable is initialized before it is
   used in a deref call.


DETAILS

    compiler/optimizations/copyPropagation.cpp
Replaced the existing implementation of copy propagation with one that tracks
aliases across the entire function.  The reference map is computed before
copy propagation begins (whether local or global), so the implementation is
somewhat conservative.  

In propagateCopies(), the RefMap is checked to see if ('deref' foo) can be
replaced by "bar" where (foo, bar) is a pair in the reference map.  The
RefMap maps reference variables to the variables they are referencing.  Since
we create the RefMap globally, we assume that the reference variable is created
and initialized before it is used in a deref call.  If that assumption is not
correct, the problem most likely lies elsewhere.

In removeKilledSymbols(), the RefMap is used to find all variables that may
have been updated through a reference, and removes all candidate pairs that
contain those variables.

Aside from that, both the local and global copy propagation algorithms are
copied straight out of Muchnick.

In copyPropagation(), added a test for FLAG_EXTERN, since extern function stubs
may contain _construct_tuple calls that are unresolved.  This is a workaround;
the stubs for extern function calls should obey all AST invariants.

Added another call to deadVariableElimination after
singleAssignmentRefPropagation().

As a useful debugging tool, debug printfs are enabled only if DEBUG_CP >= 1
*and* the filescope static "debug" is set to true.  This allows one to break in
one of the copyPropagation functions when fn->id == <your favorite function>,
and then set debug=1, to get verbose output.


    compiler/optimizations/deadCodeElimination.cpp
Improved copy propagation was allowing the "this" symbol to be bridged out of
methods that did not actually reference it.  But the AST expects every method
to have a "this" variable, so I added a special check in deadCodeElimination to
preserve it.

This case came up only in the compiler-generated default-initializer.  After
all constructors become methods and the meme operand is eliminated (future UMM
work), then this special case can be removed.


    compiler/AST/primitive.cpp
Reworked returnInfoGetMemberRef(), so it returns the right return type info if
the return value happens to be a literal.  Copy propagation is now working so
well that it can propagate literals all the way to the end of a function and
into the 'return' primitive.


    compiler/AST/symbol.cpp
Added a check that every method declares a "this" variable.


    compiler/passes/parallel.cpp
Reduced the number of variables to which FLAG_CONCURRENTLY_ACCESSED is applied,
to allow copy propagation to do a better job.  Specifically, variables can only
be concurrently accessed if they are lvalues.

The return value of a function returning a string literal is never wide.


    compiler/AST/iterator.cpp
Disabled an assert in retInitialization() because the rework of
insertYieldTemps() makes the assertion untrue. 


    compiler/include/optimizations.h
Added the DEBUG_CP precompiler symbol here, because it affects several
optimization passes (see below).


    compiler/resolution/lowerIterators.cpp
    compiler/optimizations/inlineFunctions.cpp
    compiler/optimizations/removeUnnecessaryAutoCopyCalls.cpp
Other calls to local copy propagation are disabled if DEBUG_CP >= 2.


    compiler/passes/reservedSymbolNames
Added the 'bool' keyword, because C99 uses it, and propagation removed a
promotion from bool to int (so now the type 'bool' leaks into the generated
code).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22796 3a8e244f-b0f2-452b-bcba-4c88e055c3ca