Minor cleanup to scalar replacement
[reviewed by hilde]

While working on this file this week, I noticed that the code which
rewrites expressions when scalar replacing a class or record didn't
have an } else { clause to make sure that other cases weren't slipping
between the cracks.  I added such a clause with an INT_FATAL and a
comment indicating what might be wrong if one fell into that case
(essentially, it means you added a clause saying it was OK to scalar
replace such an expression but then never added the logic saying how
to rewrite it.

While here, I also removed some redundant tests for a non-NULL pointer
which are subsumed by the enclosing conditional.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22800 3a8e244f-b0f2-452b-bcba-4c88e055c3ca